+++ partest/http/__init__.py
@@ -1,5 +1,12 @@
-"""HTTP helpers: header/param Config builder."""
+"""HTTP helpers: header/param Config builder, TLS-policy client factory."""
 
+from partest.http.client import httpx_async_client, httpx_client
 from partest.http.config import Config, HeaderSpec, HeadersBind
 
-__all__ = ["Config", "HeaderSpec", "HeadersBind"]
+__all__ = [
+    "Config",
+    "HeaderSpec",
+    "HeadersBind",
+    "httpx_client",
+    "httpx_async_client",
+]